+++ Gemfile
@@ -1,6 +1,4 @@
-# frozen_string_literal: true
-
-source "https://rubygems.org"
+source 'https://rubygems.org'
 
 gem 'rubocop', '>= 1.0', '< 2.0'
