--- modules-for-main/entry-point.rb
@@ -1,16 +1,0 @@
-class EntryPoint
-    puts 'Welcome to School Library App'
-  
-    def menu
-      puts '1 - List all books'
-      puts '2 - List all music albums'
-      puts '3 - List all games'
-      puts '4 - List all genres'
-      puts '5 - List all labels'
-      puts '6 - List all authors'
-      puts '7 - Add a book'
-      puts '8 - Add a music album'
-      puts '9 - Add a game'
-      puts '10 - Exit'
-    end
-  end
